--- a/sqlite3.go
+++ b/sqlite3.go
@@ -9,7 +9,7 @@ package sqlite3
 #cgo CFLAGS: -std=gnu99
 #cgo CFLAGS: -DSQLITE_ENABLE_RTREE -DSQLITE_THREADSAFE
 #cgo CFLAGS: -DSQLITE_ENABLE_FTS3 -DSQLITE_ENABLE_FTS3_PARENTHESIS
-#include <sqlite3-binding.h>
+#include "sqlite3-binding.h"
 #include <stdlib.h>
 #include <string.h>