… sqlite3-binding.h