The new release also conducted performance optimization of full-text queries fts3/fts4, used in combination with the **OR** operator. Fixed a potential integer overflow in functions `sqlite3_blob_read()` and `sqlite3_blob_write()`.

Before (`go test -bench=. -timeout=30s`):
```
BenchmarkExec           200000     107113 req/s
BenchmarkQuery           50000      34443 req/s
BenchmarkParams          50000      29273 req/s
BenchmarkStmt           100000      49890 req/s
BenchmarkRows             2000       1059 req/s
BenchmarkStmtRows         2000       1088 req/s
PASS
ok  	_/tmp/go/go-sqlite3-master	16.157s
```

After:
```
BenchmarkExec           200000     110548 req/s
BenchmarkQuery           50000      35105 req/s
BenchmarkParams          50000      29825 req/s
BenchmarkStmt           100000      51111 req/s
BenchmarkRows             2000       1076 req/s
BenchmarkStmtRows         2000       1100 req/s
PASS
ok  	_/tmp/go/go-sqlite3-sqlite388	15.905s
```