@mattn 2015-02-24
Thank you. But currently, golang have issue about this.

https://github.com/golang/go/issues/9356