@mattn 2016-01-23
I didn't merge this PR, but I could decide to merge latest amalgamation code to latest. thanks!
https://github.com/mattn/go-sqlite3/pull/264